extensive refactor
per link (below), refactor starts at top of page and goes until "Onwards to Part 6"

http://rogueliketutorials.com/tutorials/tcod/v2/part-6/

Per that page, here’s the basic ideas about the changes made:

Event handlers will have the handle_events method instead of Engine.
The game map will have a reference to Engine, and entities will have a reference to the map.
Actions will be initialized with the entity doing the action
Because of the above points, Actions will have a reference to the Engine, through Entity->GameMap->Engine